Ok intellij, removing the int cast

+++ core/src/mindustry/logic/LStatements.java
@@ -328,7 +328,7 @@ public class LStatements{
                         i.left();
                         int c = 0;
                         for(char j=32;j<127;j++){
-                            final int chr = (int)j;
+                            final int chr = j;
                             i.button(String.valueOf(j),Styles.flatt, () -> {
                                 value = Integer.toString(chr);
                                 field.setText(value);
